+++ argopy/tests/test_data_fetchers.py
@@ -24,8 +24,8 @@ import argopy
 from argopy import DataFetcher as ArgoDataFetcher
 from argopy.errors import InvalidFetcherAccessPoint, InvalidFetcher
 
-from argopy.utilities import list_available_data_backends
-AVAILABLE_BACKENDS = list_available_data_backends()
+from argopy.utilities import list_available_data_src
+AVAILABLE_SOURCES = list_available_data_src()
 
 import urllib.request
 def connected(host='http://www.ifremer.fr'):
@@ -45,10 +45,10 @@ def test_invalid_fetcher():
     with pytest.raises(InvalidFetcher):
         ArgoDataFetcher().to_xarray() # Can't get data if access point not defined first
 
-@unittest.skipUnless('localftp' in AVAILABLE_BACKENDS, "requires localftp data fetcher")
+@unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
 def test_unavailable_accesspoint():
     with pytest.raises(InvalidFetcherAccessPoint):
-        ArgoDataFetcher(backend='localftp').region([-85,-45,10.,20.,0,100.]).to_xarray()
+        ArgoDataFetcher(src='localftp').region([-85,-45,10.,20.,0,100.]).to_xarray()
 
 class EntryPoints_AllBackends(TestCase):
     """ Test main API facade for all available fetching backends and default dataset """
@@ -72,68 +72,68 @@ class EntryPoints_AllBackends(TestCase):
         """ Test float for a given backend """
         for arg in self.args['float']:
             options = {**self.fetcher_opts, **ftc_opts}
-            ds = ArgoDataFetcher(backend=bk, **options).float(arg).to_xarray()
+            ds = ArgoDataFetcher(src=bk, **options).float(arg).to_xarray()
             assert isinstance(ds, xr.Dataset) == True
 
     def __test_profile(self, bk):
         """ Test float for a given backend """
         for arg in self.args['profile']:
-            ds = ArgoDataFetcher(backend=bk).profile(*arg).to_xarray()
+            ds = ArgoDataFetcher(src=bk).profile(*arg).to_xarray()
             assert isinstance(ds, xr.Dataset) == True
 
     def __test_region(self, bk):
         """ Test float for a given backend """
         for arg in self.args['region']:
-            ds = ArgoDataFetcher(backend=bk).region(arg).to_xarray()
+            ds = ArgoDataFetcher(src=bk).region(arg).to_xarray()
             assert isinstance(ds, xr.Dataset) == True
 
-    @unittest.skipUnless('erddap' in AVAILABLE_BACKENDS, "requires erddap data fetcher")
+    @unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
     @unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
     def test_float_erddap(self):
         self.__test_float('erddap')
 
-    @unittest.skipUnless('erddap' in AVAILABLE_BACKENDS, "requires erddap data fetcher")
+    @unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
     @unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
     def test_profile_erddap(self):
         self.__test_profile('erddap')
 
-    @unittest.skipUnless('erddap' in AVAILABLE_BACKENDS, "requires erddap data fetcher")
+    @unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
     @unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
     def test_region_erddap(self):
         self.__test_region('erddap')
 
-    @unittest.skipUnless('localftp' in AVAILABLE_BACKENDS, "requires localftp data fetcher")
+    @unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
     def test_float_localftp(self):
         ftproot, flist = argopy.tutorial.open_dataset('localftp')
         with argopy.set_options(local_ftp=os.path.join(ftproot,'dac')):
             self.__test_float('localftp', )
 
-    @unittest.skipUnless('localftp' in AVAILABLE_BACKENDS, "requires localftp data fetcher")
+    @unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
     def test_profile_localftp(self):
         ftproot, flist = argopy.tutorial.open_dataset('localftp')
         with argopy.set_options(local_ftp=os.path.join(ftproot,'dac')):
             self.__test_profile('localftp')
 
-    @unittest.skipUnless('argovis' in AVAILABLE_BACKENDS, "requires argovis data fetcher")
+    @unittest.skipUnless('argovis' in AVAILABLE_SOURCES, "requires argovis data fetcher")
     @unittest.skipUnless(CONNECTED, "argovis requires an internet connection")
     def test_float_argovis(self):
         self.__test_float('argovis')
 
-@unittest.skipUnless('erddap' in AVAILABLE_BACKENDS, "requires erddap data fetcher")
+@unittest.skipUnless('erddap' in AVAILABLE_SOURCES, "requires erddap data fetcher")
 @unittest.skipUnless(CONNECTED, "erddap requires an internet connection")
 class Erddap_backend(TestCase):
     """ Test main API facade for all available dataset of the ERDDAP fetching backend """
 
     def test_cachepath(self):
-        assert isinstance(ArgoDataFetcher(backend='erddap').profile(6902746, 34).fetcher.cachepath, str) == True
+        assert isinstance(ArgoDataFetcher(src='erddap').profile(6902746, 34).fetcher.cachepath, str) == True
 
     def test_caching(self):
         cachedir = os.path.expanduser(os.path.join("~",".argopytest_tmp"))
         try:
             # 1st call to load from erddap and save to cachedir:
-            ds = ArgoDataFetcher(backend='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
+            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
             # 2nd call to load from cached file
-            ds = ArgoDataFetcher(backend='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
+            ds = ArgoDataFetcher(src='erddap', cache=True, cachedir=cachedir).profile(6902746, 34).to_xarray()
             assert isinstance(ds, xr.Dataset) == True            
             shutil.rmtree(cachedir)
         except:
@@ -146,31 +146,31 @@ class Erddap_backend(TestCase):
             if access_point == 'profile':
                 for arg in self.args['profile']:
                     try:
-                        ds = ArgoDataFetcher(backend='erddap', ds=dataset).profile(*arg).to_xarray()
+                        ds = ArgoDataFetcher(src='erddap', ds=dataset).profile(*arg).to_xarray()
                         assert isinstance(ds, xr.Dataset) == True
                     except:
                         print("ERDDAP request:\n",
-                              ArgoDataFetcher(backend='erddap', ds=dataset).profile(*arg).fetcher.url)
+                              ArgoDataFetcher(src='erddap', ds=dataset).profile(*arg).fetcher.url)
                         pass
 
             if access_point == 'float':
                 for arg in self.args['float']:
                     try:
-                        ds = ArgoDataFetcher(backend='erddap', ds=dataset).float(arg).to_xarray()
+                        ds = ArgoDataFetcher(src='erddap', ds=dataset).float(arg).to_xarray()
                         assert isinstance(ds, xr.Dataset) == True
                     except:
                         print("ERDDAP request:\n",
-                              ArgoDataFetcher(backend='erddap', ds=dataset).float(arg).fetcher.url)
+                              ArgoDataFetcher(src='erddap', ds=dataset).float(arg).fetcher.url)
                         pass
 
             if access_point == 'region':
                 for arg in self.args['region']:
                     try:
-                        ds = ArgoDataFetcher(backend='erddap', ds=dataset).region(arg).to_xarray()
+                        ds = ArgoDataFetcher(src='erddap', ds=dataset).region(arg).to_xarray()
                         assert isinstance(ds, xr.Dataset) == True
                     except:
                         print("ERDDAP request:\n",
-                              ArgoDataFetcher(backend='erddap', ds=dataset).region(arg).fetcher.url)
+                              ArgoDataFetcher(src='erddap', ds=dataset).region(arg).fetcher.url)
                         pass
 
     def test_phy_float(self):
@@ -215,7 +215,7 @@ class Erddap_backend(TestCase):
                                [-70, -65, 35., 40., 0, 10., '2012-01-01', '2012-12-31']]
         self.__testthis('ref')
 
-@unittest.skipUnless('localftp' in AVAILABLE_BACKENDS, "requires localftp data fetcher")
+@unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
 class LocalFTP_DataSets(TestCase):
     """ Test main API facade for all available dataset of the localftp fetching backend """
 
@@ -228,22 +228,22 @@ class LocalFTP_DataSets(TestCase):
                 for arg in self.args['profile']:
                     with argopy.set_options(local_ftp=self.local_ftp):
                         try:
-                            ds = ArgoDataFetcher(backend='localftp', ds=dataset).profile(*arg).to_xarray()
+                            ds = ArgoDataFetcher(src='localftp', ds=dataset).profile(*arg).to_xarray()
                             assert isinstance(ds, xr.Dataset) == True
                         except:
                             print("LOCALFTP request:\n",
-                                  ArgoDataFetcher(backend='localftp', ds=dataset).profile(*arg).fetcher.files)
+                                  ArgoDataFetcher(src='localftp', ds=dataset).profile(*arg).fetcher.files)
                             pass
 
             if access_point == 'float':
                 for arg in self.args['float']:
                     with argopy.set_options(local_ftp=self.local_ftp):
                         try:
-                            ds = ArgoDataFetcher(backend='localftp', ds=dataset).float(arg).to_xarray()
+                            ds = ArgoDataFetcher(src='localftp', ds=dataset).float(arg).to_xarray()
                             assert isinstance(ds, xr.Dataset) == True
                         except:
                             print("LOCALFTP request:\n",
-                                  ArgoDataFetcher(backend='localftp', ds=dataset).float(arg).fetcher.files)
+                                  ArgoDataFetcher(src='localftp', ds=dataset).float(arg).fetcher.files)
                             pass
 
     def test_phy_float(self):